Wran 5: Added separators to graphs

+++ src/com/graphs/lib/graph/element/Point.java
@@ -1,27 +1,27 @@
 package com.graphs.lib.graph.element;
 
 public class Point {
-    private int x;
-    private int y;
+    private float x;
+    private float y;
 
-    public Point(int x, int y) {
+    public Point(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
-    public int getX() {
+    public float getX() {
         return x;
     }
 
-    public void setX(int x) {
+    public void setX(float x) {
         this.x = x;
     }
 
-    public int getY() {
+    public float getY() {
         return y;
     }
 
-    public void setY(int y) {
+    public void setY(float y) {
         this.y = y;
     }
 }
